Replaced hashmap with an array list
Chosen structure (hashmap) doesnt work well with other parts of the program. Execve expects an array of pointers char ** to work with. In this case it looks counterprodactive to transform given array to a hashmap and then transform it back in order to execute some binary. Hence, replacing  hashmap with an arraylist as it has an array as underlying struct and it will be possible to expand it so it dynamically handles size management for us.

Changes:
- array list and basic opeations;
- added this to the state struct;
- updated export, unset and env to work is arraylist instead of map;
- added unit tests;